+++ src/prefect/pydantic/main.py
@@ -1,6 +1,6 @@
 import typing
 
-from prefect._internal.pydantic._compat import BaseModel
+from prefect._internal.pydantic._compat import BaseModel, Field, TypeAdapter
 
 
 class PrefectBaseModel(BaseModel):
@@ -12,4 +12,4 @@ class PrefectBaseModel(BaseModel):
         return set()
 
 
-__all__ = ["BaseModel", "PrefectBaseModel"]
+__all__ = ["BaseModel", "Field", "PrefectBaseModel", "TypeAdapter"]
